refactor: refactoring create meals logic
