+++ src/core/use-cases/factories/make-create-meal-use-case.ts
@@ -1,0 +1,9 @@
+import { MethodsMealsRepository } from "@/repositories/db-meals-repository";
+import { CreateMealUseCase } from "../meals/create-meal";
+
+export function makeCreateMealUseCase() {
+  const mealsRepository = new MethodsMealsRepository();
+  const createMealUseCase = new CreateMealUseCase(mealsRepository);
+
+  return createMealUseCase;
+}
